number of stops collected by getFiveClosestStops is now a parameter and the function is renamed to getClosestStops, also added checks to getStopsData, getStopData and getLineData to prevent some unnecessary fetch calls

+++ NaNdroid/app/src/main/java/asia/jokin/ohjelmistomobiili/MapsActivity.kt
@@ -107,7 +107,7 @@ class MapsActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnCamera
 
     private fun openStop(mCode: String) {
         val popupIntent = Intent(this@MapsActivity, PopupActivity::class.java)
-        popupIntent.putExtra("stopid", mCode)
+        popupIntent.putExtra("stopid", mCode.toInt())
         startActivity(popupIntent)
     }
 }
